Add files via upload
Hello,
My name is Vishnu Gupta and I would like to present my small project called Hospital. The purpose of this project is to create a simple application that can display the information of doctors and patients in a hospital. This project is not very complex, but it demonstrates some basic features of programming and data management. I hope you find it useful and informative. Thank you for your attention.